requirements.txt
Django>=5.0
 djangorestframework
 drf-yasg
